added arrow keys to vim keys with another layer
with the brackets on both hands I miss the arrow key vim key that the
symbol layer had, it is incredibly useful.

However it makes no sense to have it in the symbol layer.
I have made a new layer activated by double tapping the shell layer
option. I feel like I am making too many layers, plus multi tap is
sluggish.

Hopefully I can think of a better solution, but this is it for now.